Expander refactored to mostly avoid need for CPS
Imports are now analysed before the main part of the expander runs, so
all loading is done up front, the loads are cached, and then spliced in
where (import ...) forms are used. Result is that CPS is only required
for the import loading stage now, and not each of the expand
transformers.